backend/requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
decorator==4.4.2
dnspython==2.7.0
et_xmlfile==2.0.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.1.0
Flask-Cors==5.0.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
MarkupSafe==3.0.2
moviepy==2.1.2
numpy==2.2.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plyer==2.1.0
proglog==0.1.10
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
PyJWT==2.10.1
pymongo==4.11.1
pyparsing==3.2.1
pypdf==5.4.0
PyPDF2==3.0.1
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2025.1
pywin32==308
QtPy==2.4.3
requests==2.32.3
requests-oauthlib==2.0.0
resemble==1.5.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
ttkthemes==3.2.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.9.0
